changed the names of the packages in accordance with the recommendations.
