--- src/main/java/com.etorrus/factory/DepartmentFactory.java
@@ -1,4 +1,0 @@
-package com.etorrus.factory;
-
-public class DepartmentFactory extends StaffFactory{
-}
