--- src/main/java/com.etorrus/factory/OrganizationFactory.java
@@ -1,4 +1,0 @@
-package com.etorrus.factory;
-
-public class OrganizationFactory extends StaffFactory{
-}
