+++ src/main/java/com.etorrus/factory/document/DocumentFactory.java
@@ -1,4 +1,4 @@
-package com.etorrus.factory;
+package com.etorrus.factory.document;
 
 public class DocumentFactory {
