+++ src/main/java/com.etorrus/factory/document/OutgoingFactory.java
@@ -1,11 +1,9 @@
-package com.etorrus.factory;
+package com.etorrus.factory.document;
 
-import com.etorrus.documents.Outgoing;
-import com.etorrus.staffs.Person;
-import com.etorrus.util.PersonWrapper;
-import com.etorrus.dao.JaxbParser;
+import com.etorrus.model.document.Outgoing;
+import com.etorrus.model.staff.Person;
+import com.etorrus.service.PersonService;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -29,10 +27,10 @@ public class OutgoingFactory extends DocumentFactory {
         outgoing.setDateReg(new Date());
 
 
-        final JaxbParser parser = new JaxbParser();
+        /*final JaxbParser parser = new JaxbParser();
         final File file = new File("E:/xmlDocument/personList.xml");
-        final PersonWrapper personWraper = (PersonWrapper) parser.XMLToStaff(file, PersonWrapper.class);
-        final List<Person> personList = personWraper.getPersonList();
+        final PersonWrapper personWraper = (PersonWrapper) parser.XMLToStaff(file, PersonWrapper.class);*/
+        final List<Person> personList = PersonService.getPersonList();
         if (personList != null && !personList.isEmpty()) {
             Person author = personList.get(0);
             if (author != null) {
