+++ src/main/java/com.etorrus/factory/document/TaskFactory.java
@@ -1,12 +1,10 @@
-package com.etorrus.factory;
+package com.etorrus.factory.document;
 
 
-import com.etorrus.documents.Task;
-import com.etorrus.staffs.Person;
-import com.etorrus.util.PersonWrapper;
-import com.etorrus.dao.JaxbParser;
+import com.etorrus.model.document.Task;
+import com.etorrus.model.staff.Person;
+import com.etorrus.service.PersonService;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -30,10 +28,10 @@ public class TaskFactory extends DocumentFactory {
         task.setRegNum("t-00" + task.getId() + i);
         task.setDateReg(new Date());
 
-        final JaxbParser parser = new JaxbParser();
+        /*final JaxbParser parser = new JaxbParser();
         final File file = new File("E:/xmlDocument/personList.xml");
-        final PersonWrapper personWraper = (PersonWrapper) parser.XMLToStaff(file, PersonWrapper.class);
-        final List<Person> personList = personWraper.getPersonList();
+        final PersonWrapper personWraper = (PersonWrapper) parser.XMLToStaff(file, PersonWrapper.class);*/
+        final List<Person> personList = PersonService.getPersonList();
         if (personList != null && !personList.isEmpty()) {
             Person author = personList.get(0);
             if (author != null) {
