+++ src/main/java/com.etorrus/factory/staff/DepartmentFactory.java
@@ -1,0 +1,4 @@
+package com.etorrus.factory.staff;
+
+public class DepartmentFactory extends StaffFactory {
+}
