+++ src/main/java/com.etorrus/factory/staff/OrganizationFactory.java
@@ -1,0 +1,4 @@
+package com.etorrus.factory.staff;
+
+public class OrganizationFactory extends StaffFactory {
+}
