+++ src/main/java/com.etorrus/model/document/Storable.java
@@ -1,4 +1,4 @@
-package com.etorrus.documents;
+package com.etorrus.model.document;
 
 public interface Storable {
