+++ src/main/java/com.etorrus/model/document/Task.java
@@ -1,6 +1,6 @@
-package com.etorrus.documents;
+package com.etorrus.model.document;
 
-import com.etorrus.staffs.Person;
+import com.etorrus.model.staff.Person;
 
 import java.util.Date;
