+++ src/main/java/com.etorrus/model/staff/Department.java
@@ -1,16 +1,10 @@
-package com.etorrus.staffs;
+package com.etorrus.model.staff;
 
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
-/*Организация (Organization)
-полное наименование;
-краткое наименование;
-руководитель;
-контактные телефоны.
-*/
 @XmlRootElement
-public class Organization extends Staff {
+public class Department extends Staff {
     //Полное наименование
     private String fullName;
     //Краткое наименование
@@ -58,12 +52,11 @@ public class Organization extends Staff {
 
     @Override
     public String toString() {
-        return "Organization{" +
+        return "Department{" +
                 ", fullName=" + fullName + '\'' +
                 ", shortName=" + shortName + '\'' +
                 ", head=" + head + '\'' +
                 ", contactPhones=" + contactPhones +
                 '}';
     }
-
 }
