--- src/main/java/com.etorrus/staffs/PersonManager.java
@@ -1,11 +1,0 @@
-package com.etorrus.staffs;
-
-/**
- * Created by Слава on 08.12.2016.
- */
-public class PersonManager {
-
-    /*public getPersons(){
-
-    }*/
-}
