+++ src/main/java/com.etorrus/view/Report.java
@@ -1,10 +1,8 @@
 package com.etorrus.view;
 
-import com.etorrus.staffs.Person;
-import com.etorrus.util.PersonWrapper;
-import com.etorrus.dao.JaxbParser;
+import com.etorrus.model.staff.Person;
+import com.etorrus.service.PersonService;
 
-import java.io.File;
 import java.util.List;
 
 /**Отчет должен содержать перечень авторов документов
@@ -17,10 +15,10 @@ public class Report {
         System.out.println("==========REPORT==========");
 
 
-        final JaxbParser parser = new JaxbParser();
+        /*final JaxbParser parser = new JaxbParser();
         final File file = new File("E:/xmlDocument/personList.xml");
-        final PersonWrapper personWraper = (PersonWrapper) parser.XMLToStaff(file, PersonWrapper.class);
-        final List<Person> personList = personWraper.getPersonList();
+        final PersonWrapper personWraper = (PersonWrapper) parser.XMLToStaff(file, PersonWrapper.class);*/
+        final List<Person> personList = PersonService.getPersonList();
 
         for (Person author : personList){
             System.out.println(author.getLastName() + " " + author.getFirstName() + " " + author.getPatronymic());
